feat: 🎸 Add and Edit Functionality

+++ my-app/src/app/courses/course-list/course-list.component.ts
@@ -28,6 +28,8 @@ export class CourseListComponent implements OnInit {
     this.router.navigateByUrl('/courses/edit/' + index);
   }
   deleteCourse(index) {
-    this.courseItemsOverview = this.courseService.deleteCourse(index);
+    if (confirm('Are you sure to delete?')) {
+      this.courseItemsOverview = this.courseService.deleteCourse(index);
+    }
   }
 }
